backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
openai-whisper==20231117
torch
numpy
# Note: ffmpeg MUST be installed on the system path for Whisper to process .webm
pytest
pytest-asyncio
